refactor(frontend): decide template escaping with a real parser, not a hand-rolled scan
