+++ frontend/src/lib/components/apps/components/helpers/InputValue.svelte
@@ -18,12 +18,8 @@
 	import { accessPropertyByPath } from '../../utils'
 	import { computeGlobalContext, eval_like } from './eval'
 	import { deepEqual } from 'fast-equals'
-	import {
-		deepMergeWithPriority,
-		escapeTemplateBackticks,
-		isCodeInjection,
-		readFieldsRecursively
-	} from '$lib/utils'
+	import { deepMergeWithPriority, isCodeInjection, readFieldsRecursively } from '$lib/utils'
+	import { escapeTemplateBackticks } from '$lib/utils/templateLiteral'
 	import sum from 'hash-sum'
 	import { createDispatcherIfMounted } from '$lib/createDispatcherIfMounted'
