+++ frontend/src/lib/utils.test.ts
@@ -10,9 +10,7 @@ import {
 	getQueryStmtCountHeuristic,
 	isJobResolvable,
 	parseDbInputFromAssetSyntax,
-	apiErrorMessage,
-	escapeTemplateBackticks,
-	unescapeTemplateBackticks
+	apiErrorMessage
 } from './utils'
 
 // Mirrors the backend invariant that only `status = 'failure'` rows can carry a
@@ -598,93 +596,5 @@ describe('apiErrorMessage', () => {
 
 	it('handles a plain Error', () => {
 		expect(apiErrorMessage(new Error('boom'))).toBe('boom')
-	})
-})
-
-// Template mode stores its value as a JS template literal, so backticks in the text have to be
-// escaped. Escaping them inside `${...}` too is what broke nested template literals: a backslash
-// is a syntax error in expression position.
-describe('escapeTemplateBackticks', () => {
-	const nested =
-		'-p ${flow_input.iter.value}${results.pw ? ` --vault-password-file ${results.pw}` : ""} -H ${flow_input.hostname}'
-
-	it('leaves a nested template literal inside an interpolation intact', () => {
-		const expr = '`' + escapeTemplateBackticks(nested) + '`'
-		expect(expr).not.toContain('\\`')
-		expect(
-			new Function('flow_input', 'results', 'return ' + expr)(
-				{ iter: { value: 'playbook.yml' }, hostname: 'host1' },
-				{ pw: '/tmp/pw.txt' }
-			)
-		).toBe('-p playbook.yml --vault-password-file /tmp/pw.txt -H host1')
-	})
-
-	it('still escapes a backtick in the literal text', () => {
-		expect(escapeTemplateBackticks('a ` b')).toBe('a \\` b')
-		expect(escapeTemplateBackticks('a ` ${x} ` b')).toBe('a \\` ${x} \\` b')
-	})
-
-	it('leaves an escaped interpolation as literal text', () => {
-		// `\\${...}` is escaped in the template source, so the backticks inside it are literal
-		// text and still need escaping.
-		expect(escapeTemplateBackticks('\\${foo `bar`}')).toBe('\\${foo \\`bar\\`}')
-		expect(unescapeTemplateBackticks('\\${foo \\`bar\\`}')).toBe('\\${foo `bar`}')
-		expect(
-			() => new Function('return `' + escapeTemplateBackticks('\\${foo `bar`}') + '`')
-		).not.toThrow()
-	})
-
-	it('does not mistake a brace or a backtick inside a string for the end of an interpolation', () => {
-		expect(escapeTemplateBackticks('${ x["}"] } `')).toBe('${ x["}"] } \\`')
-		expect(escapeTemplateBackticks("${ f({ a: '`' }) } `")).toBe("${ f({ a: '`' }) } \\`")
-	})
-
-	it('falls back to escaping everything when the walk desynchronizes', () => {
-		// A regex literal or a comment can hide an odd quote from the walk; escaping every
-		// backtick is worse output but still parses, which a half-walked one would not.
-		for (const v of ["${ x.replace(/'/g, '') } `", '${ /* ` */ x } `']) {
-			expect(() => new Function('return `' + escapeTemplateBackticks(v) + '`')).not.toThrow()
-			expect(unescapeTemplateBackticks(escapeTemplateBackticks(v))).toBe(v)
-		}
-	})
-
-	it('does not trust a walk that desynchronizes but lands on an empty stack', () => {
-		// A `{` inside a regex is counted as an expression brace and a later literal `}` pops it,
-		// so the stack balances while the walk has lost track of what is literal text. The output
-		// still has to parse, which is what the validation catches.
-		const v = '{"match": ${/{/.test(flow_input.x)}, "literal": "`x`"}'
-		expect(() => new Function('return `' + escapeTemplateBackticks(v) + '`')).not.toThrow()
-	})
-
-	it('round-trips through unescapeTemplateBackticks', () => {
-		for (const v of [
-			nested,
-			'a ` b',
-			'${ x["}"] } `',
-			'plain',
-			'${a}${b}',
-			'\\${a}',
-			'${cond ? `a\\`b` : ""}'
-		]) {
-			expect(unescapeTemplateBackticks(escapeTemplateBackticks(v))).toBe(v)
-		}
-	})
-	// The guarantee that matters: opening a flow in the editor and saving it back must not change
-	// the stored expression, even for input the walk cannot read (neither strategy displays those
-	// perfectly — blanket unescaping corrupts an escaped backtick inside a nested template, and
-	// the walk shows the escapes literally — but neither may rewrite what is stored).
-	it('never rewrites the stored expression on a view/save cycle', () => {
-		const inputs = [
-			'{"match": ${/{/.test(flow_input.x)}, "literal": "`x`"}',
-			'${cond ? `a\\`b` : ""}',
-			"${ x.replace(/'/g, '') } `",
-			'a ` b',
-			'${ /* ` */ x } `'
-		]
-		for (const v of inputs) {
-			const stored = escapeTemplateBackticks(v)
-			expect(() => new Function('return `' + stored + '`')).not.toThrow()
-			expect(escapeTemplateBackticks(unescapeTemplateBackticks(stored))).toBe(stored)
-		}
 	})
 })
