+++ frontend/src/lib/utils.ts
@@ -1295,114 +1295,6 @@ export function cleanExpr(expr: string | undefined): string {
 		.join('\n')
 }
 
-/**
- * Walk the body of a template literal, handing the outermost literal-text regions to `onOuterText`
- * and copying everything else verbatim. `stack` tracks interpolations, braces, quoted strings and
- * nested template literals, so a `}` or a backtick inside one of those does not end the region.
- * `onOuterText` returns the replacement it emitted and how many characters it consumed, or null to
- * let the character through. `balanced` reports whether the walk ended in the outermost region;
- * it is false when the input desynchronized the stack, and the callers fall back on it.
- */
-function walkTemplateBody(
-	text: string,
-	onOuterText: (text: string, i: number) => { emit: string; consumed: number } | null
-): { out: string; balanced: boolean } {
-	const out: string[] = []
-	const stack: ('expr' | 'brace' | "'" | '"' | '`')[] = []
-	let i = 0
-	while (i < text.length) {
-		const top = stack[stack.length - 1]
-		const c = text[i]
-		if (top === "'" || top === '"' || top === '`') {
-			if (c === '\\') {
-				out.push(text.slice(i, i + 2))
-				i += 2
-				continue
-			}
-			if (c === top) {
-				stack.pop()
-			} else if (top === '`' && c === '$' && text[i + 1] === '{') {
-				out.push('${')
-				stack.push('expr')
-				i += 2
-				continue
-			}
-		} else if (top === undefined) {
-			const hit = onOuterText(text, i)
-			if (hit) {
-				out.push(hit.emit)
-				i += hit.consumed
-				continue
-			}
-			if (c === '\\') {
-				// The text is spliced straight into a template literal, so a backslash escapes
-				// the character after it: `\${` is literal text, not an interpolation.
-				out.push(text.slice(i, i + 2))
-				i += 2
-				continue
-			}
-			if (c === '$' && text[i + 1] === '{') {
-				out.push('${')
-				stack.push('expr')
-				i += 2
-				continue
-			}
-		} else {
-			if (c === '}') stack.pop()
-			else if (c === '{') stack.push('brace')
-			else if (c === "'" || c === '"' || c === '`') stack.push(c)
-		}
-		out.push(c)
-		i++
-	}
-	return { out: out.join(''), balanced: stack.length === 0 }
-}
-
-const escapeOuterBacktick = (t: string, i: number) =>
-	t[i] === '`' ? { emit: '\\`', consumed: 1 } : null
-
-const unescapeOuterBacktick = (t: string, i: number) =>
-	t[i] === '\\' && t[i + 1] === '`' ? { emit: '`', consumed: 2 } : null
-
-/** Whether `body` is a valid template-literal body. Parses only; nothing is evaluated. */
-function parsesAsTemplateBody(body: string): boolean {
-	try {
-		// eslint-disable-next-line no-new-func
-		new Function('return `' + body + '`')
-		return true
-	} catch {
-		return false
-	}
-}
-
-/**
- * Escape the backticks of `text` so it can be wrapped in a template literal, leaving `${...}`
- * interpolations alone: there a backslash is a syntax error, and a nested template literal is
- * legitimate (`${cond ? `--flag ${x}` : ''}`).
- *
- * The walk is only a heuristic — it has no notion of regex literals or comments, and a `{` inside
- * one can leave it desynchronized yet ending on an empty stack. So its result is used only when it
- * parses and inverts; anything else falls back to escaping every backtick, which is what this did
- * before nested template literals were supported and is never worse than that.
- */
-export function escapeTemplateBackticks(text: string): string {
-	const { out, balanced } = walkTemplateBody(text, escapeOuterBacktick)
-	const trustworthy =
-		balanced &&
-		parsesAsTemplateBody(out) &&
-		walkTemplateBody(out, unescapeOuterBacktick).out === text
-	return trustworthy ? out : text.replaceAll('`', '\\`')
-}
-
-/** Inverse of {@link escapeTemplateBackticks}, for turning an expression back into a template. */
-export function unescapeTemplateBackticks(text: string): string {
-	const { out, balanced } = walkTemplateBody(text, unescapeOuterBacktick)
-	// Same heuristic, so the same guard: keep the walk only when re-escaping reproduces the
-	// input, or the pair would not round-trip through the editor.
-	const trustworthy = balanced && walkTemplateBody(out, escapeOuterBacktick).out === text
-	return trustworthy ? out : text.replaceAll('\\`', '`')
-}
-
 const dynamicTemplateRegex = new RegExp(/\$\{(.*)\}/)
 
 export function isCodeInjection(expr: string | undefined): boolean {
